Added UI and spawner
The main menu and pause menu have been added, along with the score display. The spawner has also been added, to spawn enemies and asteroids. So far, the integration from the main game to the main menu is bugged, but it will be taken care of soon, along with the inability to pause again.