Fixed pause menu, but couldn't fix scene integration
Unfortunately, I could never figure out what is going on for the scene change from the MainGame to the MainMenu. I checked the build settings, which both scenes can be found, yet, the game has trouble finding the Instance of the MainMenu scene. This is odd, since the integration from the MainMenu to the MainGame works fine, but cannot work the other way around. The coding and setup for both of the scenes' UI is essentially the same, so all of this leaves me stumped. At the very least, all of the other components seem to work perfectly.