SimonSays complete
I have finished SimonSays, but it is not completely functional. I could not figure out how to get player input into the playerButtonPressed list (the expected behaviour would be that the player pressing the button would add its id to the list). due to this, I could not properly test other items. I tried a variety of methods, but none seem to work, without having to change the structure of the SimonSays buttons entirely (UI to transformation, which would defeat the purpose of the project). Menu directories work fine, and button audio should be good, and the Simon AI works good as expected. As far as I am aware, there should not be any runtime errors.