Added dirt, UI and progress on main game functions
Dirt functions should now be complete, and I added UI to the game (mainly in the test level for now). I still need to work on successfully implementing a timer before I could implement all functions to the real levels. Meanwhile, I have also done work on sculpting the Mall level. I am not entirely sure I can figure out how to get randomization to work in time, so there is a strong possibility that it won't make it in the final build. Work on randomization will have to wait after the main functions (powers, dirt, timer, UI, and level integration) is fully ready and implemented in the Mall and MultilevelParkinglot levels. In the event that I can't put in randomization in time, I can make an "unused" folder with the unfinished scripts I have for randomization.