Work on the second level
Progress was made in the second level. Code was integrated successfully, and there were no sign of errors. The level at this point in time is almost complete, so the game project should be finished by next morning. I put the unused coding in a folder called "unused", to check for coding that was attempted for randomization. Once the game is complete, I will clean external material that I didn't use, and see if I can make a build of the game.